client: companyPage is done and working, finishing on style. server: changed mapper and added methods to companycontroller
