+++ backend/src/main/java/src/springboot/controllers/CompanyController.java
@@ -5,9 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import src.springboot.dto.NewCouponRequest;
 import src.springboot.dto.NewLoginRequest;
-import src.springboot.entities.Category;
-import src.springboot.entities.ClientType;
-import src.springboot.entities.Coupon;
+import src.springboot.entities.*;
 import src.springboot.exceptions.UnAuthorizedException;
 import src.springboot.mapper.Mapper;
 import src.springboot.service.CompanyService;
@@ -20,6 +18,8 @@ import java.util.List;
 public class CompanyController extends ClientController{
     @Autowired
     private CompanyService companyService;
+    @Autowired
+    private Mapper mapper;
 
     @Override
     @PostMapping("/login")
@@ -33,14 +33,15 @@ public class CompanyController extends ClientController{
 
     @PostMapping("/{companyId}/coupons")
     public Coupon addCoupon(@PathVariable Long companyId, @RequestBody NewCouponRequest newCouponRequest) throws UnAuthorizedException {
-        Coupon coupon = Mapper.mapToCoupon(newCouponRequest);
+        newCouponRequest.setCompanyId(companyId);
+        Coupon coupon = mapper.mapToCoupon(newCouponRequest);
         coupon.setCompany(companyService.getCompanyDetails(companyId));
         return companyService.addCoupon(coupon);
     }
 
     @PutMapping("/{companyId}/coupons")
     public Coupon updateCoupon(@PathVariable Long companyId, @RequestBody NewCouponRequest newCouponRequest) throws UnAuthorizedException {
-        Coupon coupon = Mapper.mapToCoupon(newCouponRequest);
+        Coupon coupon = mapper.mapToCoupon(newCouponRequest);
         coupon.setCompany(companyService.getCompanyDetails(companyId));
         return companyService.updateCoupon(coupon);
     }
@@ -58,6 +59,11 @@ public class CompanyController extends ClientController{
     @GetMapping("/{companyId}/coupons/price/{maxPrice}")
     public List<Coupon> getCompanyCoupons(@PathVariable Long companyId, @PathVariable double maxPrice) throws UnAuthorizedException {
         return companyService.getCompanyCoupons(companyId, maxPrice);
+    }
+
+    @GetMapping("/{companyId}")
+    public Company getCompanyDetails(@PathVariable Long companyId) throws UnAuthorizedException {
+        return companyService.getCompanyDetails(companyId);
     }
 
 }
